Completed the program by adding the final line of code that combined all previous entities to be able to produce one standard line of code that would generate a "New Band Name" in this case.
I had to give my input features a variable to be able to hold the data in which I wanted it to output there for I used the variables (city and pet).

Lastly, when I was able to combine everything together, I was able to print a standard new phrase and add the variables from the input sections that indicated what exactly I wanted the output to be.